requirements.txt
Flask==3.0.0
flask-socketio==5.3.6
flask-cors==4.0.0
python-socketio==5.10.0
python-engineio==4.7.1
SQLAlchemy==2.0.23
PyMySQL==1.1.0
python-dotenv==1.0.0
cryptography==41.0.7
eventlet==0.33.3